+++ doc/changelog.rst
@@ -50,6 +50,7 @@ Admin
 ~~~~~
 
 - Add ``admin_contact_group_form_part`` provider for ``ContactGroup`` admin
+- Redo shipping and payment method management
 - Add service provider management
 - Add package mode for products
 - Enable ```merchant_notes`` editing for ``Contact``